refactor: #8 初期ページのディレクトリ構成を変更。
