+++ bear-clone/src/app/default/pages/default-list/selected/selected.component.spec.ts
